requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
redis==5.0.4
python-dotenv==1.0.1
pytest==8.3.4
pytest-asyncio==0.22.0
httpx==0.28.1
python-multipart==0.0.20
